requirements.txt
psycopg2-binary==2.8.6
flask==1.1.2
gunicorn==20.0.4
flask-simple-geoip==0.2.3
mysql-connector-python==8.0.21
